requirements.txt
python-telegram-bot==21.6
playwright==1.48.0
python-dotenv==1.0.1
pdfplumber==0.11.4
